+++ rapyer/errors/base.py
@@ -8,3 +8,23 @@ class KeyNotFound(RapyerError):
     """Raised when a key is not found in Redis."""
 
     pass
+
+
+class FindError(RapyerError):
+    """Raised when a model cannot be found."""
+
+    pass
+
+
+class BadFilterError(FindError):
+    """Raised when a filter is invalid."""
+
+    pass
+
+
+class UnsupportedIndexedFieldError(FindError):
+    pass
+
+
+class CantSerializeRedisValueError(RapyerError):
+    pass