+++ rapyer/types/byte.py
@@ -2,6 +2,7 @@ from typing import TypeAlias, TYPE_CHECKING
 
 from pydantic_core import core_schema
 from pydantic_core.core_schema import ValidationInfo, SerializationInfo
+
 from rapyer.types.base import RedisType, REDIS_DUMP_FLAG_NAME
 
 
@@ -14,7 +15,9 @@ class RedisBytes(bytes, RedisType):
     def __iadd__(self, other):
         new_value = self + other
         if self.pipeline:
-            self.pipeline.json().set(self.key, self.json_path, new_value)
+            self.pipeline.json().set(
+                self.key, self.json_path, self.serialize_unknown(new_value)
+            )
         return self.__class__(new_value)
 
     @classmethod
@@ -52,4 +55,4 @@ class RedisBytes(bytes, RedisType):
 
 
 if TYPE_CHECKING:
-    RedisBytes: TypeAlias = RedisBytes | bytes
+    RedisBytes: TypeAlias = RedisBytes | bytes  # pragma: no cover